+++ netbox_ssl/api/serializers/__init__.py
@@ -1,8 +1,10 @@
 from .assignments import CertificateAssignmentSerializer
+from .certificate_authorities import CertificateAuthoritySerializer
 from .certificates import CertificateImportSerializer, CertificateSerializer
 
 __all__ = [
     "CertificateSerializer",
     "CertificateImportSerializer",
     "CertificateAssignmentSerializer",
+    "CertificateAuthoritySerializer",
 ]