+++ netbox_ssl/filtersets/__init__.py
@@ -1,7 +1,9 @@
 from .assignments import CertificateAssignmentFilterSet
+from .certificate_authorities import CertificateAuthorityFilterSet
 from .certificates import CertificateFilterSet
 
 __all__ = [
     "CertificateFilterSet",
     "CertificateAssignmentFilterSet",
+    "CertificateAuthorityFilterSet",
 ]