+++ netbox_ssl/models/__init__.py
@@ -1,4 +1,9 @@
 from .assignments import CertificateAssignment
+from .certificate_authorities import (
+    CATypeChoices,
+    CertificateAuthority,
+    DEFAULT_CERTIFICATE_AUTHORITIES,
+)
 from .certificates import Certificate, CertificateAlgorithmChoices, CertificateStatusChoices
 
 __all__ = [
@@ -6,4 +11,7 @@ __all__ = [
     "CertificateStatusChoices",
     "CertificateAlgorithmChoices",
     "CertificateAssignment",
+    "CertificateAuthority",
+    "CATypeChoices",
+    "DEFAULT_CERTIFICATE_AUTHORITIES",
 ]