+++ netbox_ssl/tables/__init__.py
@@ -1,7 +1,9 @@
 from .assignments import CertificateAssignmentTable
+from .certificate_authorities import CertificateAuthorityTable
 from .certificates import CertificateTable
 
 __all__ = [
     "CertificateTable",
     "CertificateAssignmentTable",
+    "CertificateAuthorityTable",
 ]